backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
langdetect==1.0.9
google-cloud-translate==3.14.0
